ft_strlen size_t

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: tspoof <tspoof@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/25 16:41:02 by tspoof            #+#    #+#             */
-/*   Updated: 2022/10/25 16:49:33 by tspoof           ###   ########.fr       */
+/*   Updated: 2022/10/26 13:04:55 by tspoof           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include <stdio.h>
+
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
